pyproject.toml
[project]
name = "personal-dev-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
